Remove data urls in favor of determining everything from slugs
This commit updates the analytics reporter so that agencies no longer need to specify a `data` property and a `slug` property. Now only a `slug` property needs to be specified since imminent changes on the reporter will make these values equal.

This changes prepares analytics.usa.gov to respond to the changes in 18f/analytics-reporter#181.