Setup continuous deploy (#436)
This commit sets up continuous deploy with Travis. Whenever a changeset is merged into master, Travis will deploy the changes.

At this time, this commit deploys production and staging simultaneously. The deploy can be configured to have a branch that deploys master and a branch that deploys staging, but that seems like a bit much for a static site.