Update production Docker image to use custom nginx config
This commit updates the Dockerfile for the production Docker image so that is uses the nginx config in the repo.

The nginx config had to be modified a bit to make this possible. Specifically, an environment variable named `S3_BUCKET_URL`. This is the url for the data in the S3 bucket. This is done so the Docker image can handle this proxying action itself, instead of needing an nginx or apache config outside the container to handle this.

Since the nginx conf depends on the environment, it needs to be built at runtime with erb meaning ruby needed to be added to the container. Since ruby was added to the container, I went ahead and moved the site build into the container as well.