[GH-1264] Add historical-data-download page
Create a react component with a form which takes inputs from the form to create
an API route to retrieve DAP API data. On submit the component makes the API calls
to retrieve the data and compiles it if there are multiple pages of data. The
component then downloads the data to the user's browser.

Add date-fns package for date manipulation and formatting
Add export-to-csv package for CSV formatting
Add reverse proxy for DAP API requests with auth header
Disable IPv6 DNS resolution for reverse proxy requests
Make the touchpoints feedback button accessible